refactor(home): fetching de datos con `useSWRInmutable`
